Refactored group tests
