+++ manager_test.go
@@ -68,16 +68,3 @@ func TestSuccessLogout(t *testing.T) {
 		})
 	})
 }
-
-// ********************* List *********************
-
-func TestListGroups(t *testing.T) {
-	convey.Convey("Given I get all groups", t, func() {
-		server := mockRequest("/clients/", "GET", 200, ``)
-		m := Manager{URL: server.URL}
-		_, err := m.ListGroups("token")
-		convey.Convey("Then It does not fail", func() {
-			convey.So(err, convey.ShouldBeNil)
-		})
-	})
-}
